+++ linter/pylintrc_recipe
@@ -1,7 +1,7 @@
 [MASTER]
 load-plugins=linter.conanv2_transition,
              linter.transform_conanfile,
-             linter.transform_import_tools
+             linter.transform_imports
 
 py-version=3.6
 recursive=no
