+++ linter/pylintrc_testpackage
@@ -1,6 +1,7 @@
 [MASTER]
 load-plugins=linter.conanv2_transition,
-             linter.conanfile_transform
+             linter.transform_conanfile,
+             linter.transform_imports
 py-version=3.6
 recursive=no
 suggestion-mode=yes
